fix some stuff in update script

+++ update_kernel_mbp.sh
@@ -75,13 +75,17 @@ done
 sed -i "s:^GRUB_CMDLINE_LINUX=.*:GRUB_CMDLINE_LINUX=\"${GRUB_CMDLINE_VALUE}\":g" /etc/default/grub
 sed -i '/^GRUB_ENABLE_BLSCFG=true/c\GRUB_ENABLE_BLSCFG=false' /etc/default/grub
 
-echo >&2 "===]> Info: Adding fedora-mbp yum repo gpg key...";
-curl -sSL "https://raw.githubusercontent.com/mikeeq/mbp-fedora-kernel/${UPDATE_SCRIPT_BRANCH}/yum-repo/fedora-mbp.gpg" > ./fedora-mbp.gpg
-rpm --import ./fedora-mbp.gpg
-rm -rfv ./fedora-mbp.gpg
+if rpm -q gpg-pubkey --qf '%{SUMMARY}\n' | grep -q -i mbp-fedora; then
+  echo >&2 "===]> Info: fedora-mbp yum repo gpg key is already added, skipping...";
+else
+  echo >&2 "===]> Info: Adding fedora-mbp yum repo gpg key...";
+  curl -sSL "https://raw.githubusercontent.com/mikeeq/mbp-fedora-kernel/${UPDATE_SCRIPT_BRANCH}/yum-repo/fedora-mbp.gpg" > ./fedora-mbp.gpg
+  rpm --import ./fedora-mbp.gpg
+  rm -rfv ./fedora-mbp.gpg
+fi
 
 echo >&2 "===]> Info: Installing kernel version: ${MBP_KERNEL_TAG}";
-rpm --force -i ./*.rpm
+rpm --force -i $(ls | grep -v -i 'mbp-fedora-t2-config')
 
 ### Suspend fix
 echo >&2 "===]> Info: Adding suspend fix... ";
